src/data/blockchain_listener.py
import asyncio
from typing import Callable, List
from solana.rpc.async_api import AsyncClient
from solana.rpc.commitment import Confirmed

class BlockchainListener:
    def __init__(self, rpc_url: str, callback: Callable):
        self.client = AsyncClient(rpc_url, commitment=Confirmed)
        self.callback = callback
        self.running = False
        
    async def start_monitoring(self):
        """
        Start monitoring the blockchain for new token creations
        """
        self.running = True
        while self.running:
            try:
                # Get latest block
                block = await self.client.get_recent_blockhash()
                
                # Look for token program interactions
                signatures = await self.client.get_signatures_for_address(
                    "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"  # Solana Token Program
                )
                
                for sig in signatures:
                    tx = await self.client.get_transaction(sig.signature)
                    if self._is_token_creation(tx):
                        token_data = self._extract_token_data(tx)
                        await self.callback(token_data)
                        
                await asyncio.sleep(1)  # Avoid rate limiting
                
            except Exception as e:
                print(f"Error monitoring blockchain: {e}")
                await asyncio.sleep(5)  # Back off on error
                
    def _is_token_creation(self, transaction: dict) -> bool:
        """
        Check if transaction is a token creation
        """
        # Implementation depends on specific blockchain
        # Need to analyze transaction logs/instructions
        pass
        
    def _extract_token_data(self, transaction: dict) -> dict:
        """
        Extract relevant token data from creation transaction
        """
        # Implementation depends on specific blockchain
        pass 