requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
numpy>=1.26.4
scipy>=1.12.0
scikit-learn>=1.4.0
aiohttp>=3.9.3
textblob>=0.17.1
python-binance>=1.0.19
pandas>=2.2.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.26.0
Pillow>=10.2.0
matplotlib>=3.8.0
base58>=2.1.1